results_loader/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
polars==1.37.1
polars-runtime-32==1.37.1
psycopg2-binary==2.9.11
requests==2.32.5
urllib3==2.6.3
